reproducible FAT order, tool hardening, reseal
- run.sh/distfiles.sh: $SORT resolved before PATH= so the FAT find
  loops and plain-dir overlay tar are host-independent (LC_ALL=C);
  collapse per-ARCH case subshells; replace ./match with test;
  FORCE_REBUILD escape hatch; loud notice when sha256sum_k0_answers
  seals vs verifies
- tools: dd.c int32->long file_offset; fatput.c bound alloc_cluster +
  loud exhaustion; bh0x.c drop redundant pre-header read
- CI: fail-fast:false on all matrix jobs; matrix.qemu replaces the arch
  ternary; action_get_host_artifact chmod via find -exec
- delete 14 orphan scripts/*.kaem + tools/diskwrite.c; rename
  boot-uefi*.sh -> boot-uefi-riscv64*.sh
- re-seal all 7 k0-*.answers (post FAT-sort fix + builder-hex0
  riscv64-stage2.hex0 regeneration)

+++ tools/dd.c
@@ -98,7 +98,9 @@ int main(int argc, char** argv) {
     }
     free(buf);
   } else {
-    int32_t file_offset = lseek(input_fd, 0, 2);
+    /* lseek returns off_t (64-bit on rv64/x86_64); use long so files
+     * >2 GiB don't silently corrupt pad_total via int32_t truncation. */
+    long file_offset = lseek(input_fd, 0, 2);
     if (file_offset == -1) {
       fputs("Error: Failed to seek to end of input file\n", stderr);
       close(input_fd);
